+++ src/main/java/bsise/server/clova/client/ClovaFeignClient.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
 
-@FeignClient(name = "clova-service", url = "${feign.clova.url}")
+@FeignClient(name = "clova-service", url = "${feign.clova.url}", configuration = ClovaFeignConfig.class)
 public interface ClovaFeignClient {
 
     @PostMapping(consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
